requirements.txt
alembic==1.0.0
aniso8601==3.0.2
Click==7.0
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.6
marshmallow-sqlalchemy==0.14.1
pkg-resources==0.0.0
psycopg2==2.7.5
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
